pyproject.toml
[project]
name = "pattern-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.0.1",
    "psycopg[async,binary,pool]>=3.3.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "yoyo-migrations>=9.0.0",
    "dishka>=1.2.0",
    "prometheus-client>=0.21.0",
    "aiohttp>=3.10.0",
]

[dependency-groups]
formatters = [
    "autoflake>=1.4,<2",
    "isort>=5.10.1,<6",
    "docformatter>=1.4,<2",
    "add-trailing-comma>=2.2.1,<3",
    "black==24.3.0",
    "ruff>=0.6.0",
    "bandit>=1.7.9",
    "safety>=3.2.0",
]

[tool.uv]
default-groups = ["formatters"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
